Restore stubs into original plan folders
